+++ hs-bindgen/test/th/Test/TH/RecordDot.hs
@@ -37,7 +37,7 @@ import HsBindgen.TH
 
 let cfg :: Config
     cfg = def
-      & #clang % #extraIncludeDirs .~ [Pkg ("test-artefacts" </> "headers")]
+      & #clang % #extraIncludeDirs .~ [PkgDir ("test-artefacts" </> "headers")]
       & #fieldNamingStrategy       .~ OmitFieldPrefixes
     cfgTh :: ConfigTH
     cfgTh = def